+++ components/shapediver/viewport/ViewportIcons.tsx
@@ -1,4 +1,8 @@
 import {OverlayPosition} from "@AppBuilderShared/components/shapediver/ui/OverlayWrapper";
+import {
+	ComponentContext,
+	renderViewportIcons,
+} from "@AppBuilderShared/context/ComponentContext";
 import {useViewportControls} from "@AppBuilderShared/hooks/shapediver/viewer/useViewportControls";
 import {useViewportId} from "@AppBuilderShared/hooks/shapediver/viewer/useViewportId";
 import {useShapeDiverStoreParameters} from "@AppBuilderShared/store/useShapeDiverStoreParameters";
@@ -10,20 +14,9 @@ import {
 	ViewportIconsOptionalProps,
 	ViewportIconsProps,
 } from "@AppBuilderShared/types/shapediver/viewportIcons";
-import {
-	Divider,
-	DividerProps,
-	Paper,
-	Transition,
-	useProps,
-} from "@mantine/core";
-import React, {useCallback, useMemo, useState} from "react";
+import {Paper, Transition, useProps} from "@mantine/core";
+import React, {useCallback, useContext, useMemo, useState} from "react";
 import {useShallow} from "zustand/react/shallow";
-import {ViewportTypeToIcon} from "~/shared/context/ComponentContext";
-import {
-	ViewportIconButtonEnum,
-	ViewportIconLayoutItem,
-} from "~/shared/types/store/shapediverStoreViewportIcons";
 import ViewportOverlayWrapper from "./ViewportOverlayWrapper";
 
 export const defaultStyleProps: ViewportIconsOptionalProps = {
@@ -57,162 +50,6 @@ export const defaultStyleProps: ViewportIconsOptionalProps = {
 	},
 };
 
-const renderButtonByKind = (kind: ViewportIconButtonEnum, context: any) => {
-	const {
-		viewport,
-		namespace,
-		buttonsDisabled,
-		executing,
-		hasPendingChanges,
-		iconsVisible,
-		fullscreenId,
-		...commonProps
-	} = context;
-
-	switch (kind) {
-		case ViewportIconButtonEnum.Ar:
-			return React.createElement(
-				ViewportTypeToIcon[ViewportIconButtonEnum.Ar],
-				{key: "ar", viewport, ...commonProps},
-			);
-		case ViewportIconButtonEnum.Zoom:
-			return React.createElement(
-				ViewportTypeToIcon[ViewportIconButtonEnum.Zoom],
-				{key: "zoom", viewport, ...commonProps},
-			);
-		case ViewportIconButtonEnum.Fullscreen:
-			return React.createElement(
-				ViewportTypeToIcon[ViewportIconButtonEnum.Fullscreen],
-				{
-					key: "fullscreen",
-					fullscreenId,
-					enableFullscreenBtn: true,
-					...commonProps,
-				},
-			);
-		case ViewportIconButtonEnum.Cameras:
-			return React.createElement(
-				ViewportTypeToIcon[ViewportIconButtonEnum.Cameras],
-				{
-					key: "cameras",
-					viewport,
-					visible: iconsVisible,
-					...commonProps,
-				},
-			);
-		case ViewportIconButtonEnum.Undo:
-			return React.createElement(
-				ViewportTypeToIcon[ViewportIconButtonEnum.Undo],
-				{
-					key: "undo",
-					disabled: buttonsDisabled || executing || hasPendingChanges,
-					hasPendingChanges,
-					executing,
-					...commonProps,
-				},
-			);
-		case ViewportIconButtonEnum.Redo:
-			return React.createElement(
-				ViewportTypeToIcon[ViewportIconButtonEnum.Redo],
-				{
-					key: "redo",
-					disabled: buttonsDisabled || executing || hasPendingChanges,
-					hasPendingChanges,
-					executing,
-					...commonProps,
-				},
-			);
-		case ViewportIconButtonEnum.Reload:
-			return React.createElement(
-				ViewportTypeToIcon[ViewportIconButtonEnum.Reload],
-				{
-					key: "reload",
-					disabled:
-						!namespace ||
-						buttonsDisabled ||
-						executing ||
-						hasPendingChanges,
-					namespace: namespace || "",
-					hasPendingChanges,
-					executing,
-					...commonProps,
-				},
-			);
-		case ViewportIconButtonEnum.HistoryMenu:
-			return React.createElement(
-				ViewportTypeToIcon[ViewportIconButtonEnum.HistoryMenu],
-				{
-					key: "historyMenu",
-					disabled:
-						!namespace || buttonsDisabled || hasPendingChanges,
-					namespace: namespace || "",
-					visible: iconsVisible,
-					...commonProps,
-				},
-			);
-		default:
-			return null;
-	}
-};
-
-const renderViewportIcons = (
-	viewportIcons: ViewportIconLayoutItem[],
-	buttonContext: ButtonRenderContext,
-	dividerProps: DividerProps = {},
-) => {
-	const sections: React.ReactNode[] = [];
-	if (viewportIcons.length === 0) return sections;
-
-	viewportIcons.forEach((item, index) => {
-		if (item.type === "button") {
-			const button = renderButtonByKind(item.button.type, buttonContext);
-			if (button) sections.push(button);
-		} else if (item.type === "group") {
-			const groupButtons: React.ReactNode[] = [];
-			item.sections.forEach((section) => {
-				section.forEach((buttonDef) => {
-					const button = renderButtonByKind(
-						buttonDef.type,
-						buttonContext,
-					);
-					if (button) groupButtons.push(button);
-				});
-				// Add divider between sections within a group
-				if (
-					groupButtons.length > 0 &&
-					section !== item.sections[item.sections.length - 1]
-				) {
-					groupButtons.push(
-						React.createElement(Divider, {
-							key: `divider-${index}-${section.length}`,
-							...dividerProps,
-						}),
-					);
-				}
-			});
-			sections.push(
-				React.createElement(
-					React.Fragment,
-					{key: `group-${index}`},
-					...groupButtons,
-				),
-			);
-		}
-
-		// Add divider between layout items
-		if (index < viewportIcons.length - 1) {
-			sections.push(
-				React.createElement(Divider, {
-					key: `layout-divider-${index}`,
-					...dividerProps,
-				}),
-			);
-		}
-	});
-
-	return sections;
-};
-
 export default function ViewportIcons(
 	props: ViewportIconsProps & ViewportIconsOptionalProps,
 ) {
@@ -223,6 +60,7 @@ export default function ViewportIcons(
 		...rest
 	} = props;
 
+	const componentContext = useContext(ComponentContext);
 	const {
 		style,
 		iconStyle,
@@ -358,8 +196,13 @@ export default function ViewportIcons(
 	);
 
 	const content = useMemo(() => {
-		return renderViewportIcons(layout, buttonContext, dividerProps);
-	}, [layout, buttonContext, dividerProps]);
+		return renderViewportIcons(
+			layout,
+			componentContext,
+			buttonContext,
+			dividerProps,
+		);
+	}, [layout, componentContext, buttonContext, dividerProps]);
 
 	// Prevent event propagation to avoid triggering viewport interactions
 	// when touching the icons.
